Norminette compliant environment
