+++ Makefile
@@ -7,7 +7,7 @@ TOK_DIR = src/parsing/tokenisation
 AST_DIR = src/parsing/ast
 ENV_DIR = src/parsing/environment
 UTILS_DIR = src/utils
-MINI_ENV_SRC = $(ENV_DIR)/env_node.c $(ENV_DIR)/env_list.c $(ENV_DIR)/env_utils.c
+MINI_ENV_SRC = $(ENV_DIR)/env_node.c $(ENV_DIR)/env_list.c $(ENV_DIR)/env_list_1.c $(ENV_DIR)/env_utils.c
 MINI_UTILS_SRC = $(UTILS_DIR)/utils.c
 MINI_AST_SRC = $(AST_DIR)/parse.c
 MINI_TOKENISATION_SRC = $(TOK_DIR)/tokens.c $(TOK_DIR)/single_token_utils.c $(TOK_DIR)/token_list_utils.c $(TOK_DIR)/tokens_handlers.c $(TOK_DIR)/tokens_handler_01.c
