+++ src/parsing/environment/env_list.c
@@ -6,7 +6,7 @@
 /*   By: jnenczak <jnenczak@student.42roma.it>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/18 18:14:46 by jnenczak          #+#    #+#             */
-/*   Updated: 2024/11/19 20:14:21 by jnenczak         ###   ########.fr       */
+/*   Updated: 2024/11/19 20:16:31 by jnenczak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,63 +41,6 @@ t_env_list	*env_list_initialize(void)
 	ret->head = NULL;
 	ret->tail = NULL;
 	return (ret);
-}
-
-void	_env_list_append_node(t_env_list *list, t_env_node *node)
-{
-	list->tail->next = node;
-	node->prev = list->tail;
-	list->tail = node;
-}
-
-int	_env_list_add_head(t_env_list *list, t_env_node *node)
-{
-	if (list->head == NULL && list->tail == NULL)
-	{
-		list->head = node;
-		list->tail = node;
-		return (1);
-	}
-	return (0);
-}
-
-void	_env_list_place_node(t_env_list *list, t_env_node *current, t_env_node *node)
-{
-	node->next = current;
-	node->prev = current->prev;
-	if (current->prev)
-		current->prev->next = node;
-	else
-		list->head = node;
-	current->prev = node;
-}
-
-void	env_list_insert_node(t_env_list *list, t_env_node *node)
-{
-	t_env_node	*current;
-	int			comp_res;
-
-	if (!list || !node)
-		return ;
-	if (_env_list_add_head(list, node))
-		return ;
-	current = list->head;
-	while (current != NULL)
-	{
-		if (current->next == NULL)
-			break ;
-		comp_res = env_node_comp(current, node->key);
-		if (comp_res <= 0)
-		{
-			if (comp_res < 0)
-				_env_list_place_node(list, current, node);
-			else
-				env_node_update(current, node->value);
-			return ;
-		}
-		current = current->next;
-	}
-	_env_list_append_node(list, node);
 }
 
 t_env_node	*env_list_read_node(t_env_list *list, char *key)
@@ -159,22 +102,4 @@ void	env_list_delete(t_env_list *list)
 	temp = NULL;
 	free(list);
 	list = NULL;
-}
-
-void	env_list_insert(t_env_list *list, const char *entry)
-{
-	t_env_node	*node;
-
-	if (!list || !entry)
-	{
-		log_error("[ENV_LIST_INSERT] Failed to initialize a new node.");
-		return ;
-	}
-	node = env_node_initialize(entry);
-	if (!node)
-	{
-		log_error("[ENV_LIST_INSERT] Failed to initialize a new node.");
-		return ;
-	}
-	env_list_insert_node(list, node);
 }
